+++ src/main/java/edu/icet/hospital_system/repository/custom/PaymentDao.java
@@ -4,5 +4,5 @@ import edu.icet.hospital_system.entity.PaymentEntity;
 import edu.icet.hospital_system.repository.CrudDao;
 
 public interface PaymentDao extends CrudDao<PaymentEntity,String> {
-    Long getTotIncome();
+    Double getTotIncome();
 }
